Set up project for TypeScript and start porting
After a bit of trial an error, porting has commenced. The classes
that are "finished" thus far are Actor, Entity, Point and Scene.
Additionally the core namespace that contains things like the stage
width and the enumeration for key constants is also ported.

There is currently a partial implementation of the Stage class.
Additionally there are some TODO comments scattered in the ported
code to remind me to go back and tweak those areas more.