Finish the initial port of the Stage class
This can use some work. In particular, the original JavaScript
version was using functions caught in an IIFE to do things like input
and scene switching.

These are currently static class methods, which is the closest
approximation to that. However the original intention was to hide the
methods, which can be done now via member hiding, so that needs to be
looked at.