Finish the initial port of the Tile class
This port is complete but I've created an issue in the repository to
address modifying the API for whether or not this tile blocks actor
movement to take an actual actor, allowing tiles to block some actors
but not others.

This occurred to me when working on entities last night, which should
be modified the same way.