Finish the initial port of the LevelData class
As is often the case in situations like this, the API of the Actor
and Entity classes needed to be tweaked to allow for the required
member access.

I'm not 100% sure I like the way that I've gone about this. There's
an issue in the repository on gitlab to investigate this more. In
particular I am not a fan of warty members (e.g. _properties) but I
also want the accessor methods to have a decent name, which I've
already chosen for the properties itself.

Also, there may or may not be a speed issue. I'm more hung up on the
aesthetics, currently.