Finish the initial porting effort of all classes
There are still a few TODO items to clean up and things to look at
before this stage is complete, but so far so good. He said, having
not actually tested anything as of yet.

The last of the initial steps will be to port over the mainline code
that the page uses as a template to set things up, and then we can
modify the HTML page to include the new single generated file and see
what happens.