Initial porting efforts are now complete
The initial sample (which admittedly does a whole lot of nothing) is
now completely implemented in TypeScript. The original JS version
will be kept around for a bit longer in this repository.

In porting over the contentLoaded code I was originally using, I
discovered that I could in fact create functions and not just methods.
I just assumed (shame on me) that only classes could be created,
because that's how I'm used to languages like this working.

This will make fixing up some of the TODO items in the Stage class
easier, I think.