Amend the entity search API in the Scene class
Per the issue in the issue tracker, the original API for finding
entities was created before I decided that all Actors and not just
specific Entity subclasses should have a map position AND a world
position maintained simultaneously.

As such, the API is extended to be able to find generic actors at a
map position, which is probably more interesting in the general case.
A side benefit of this is that since the map position is always
updated when the world position changes, this makes it easier to find
Actors in a general area, which is nice.