Make the Tileset class validate tilesets better
This resolves the issue where no warning would be generated if a tile
with a duplicate textual name or numeric ID was given to the tileset.

This is now handled by dumping a warning out to the console, since
this is primarily a design time issue and there's no graceful way to
handle the situation without bombing the entire engine out.