Port over the JavaScript String.format function from the old code
Technically this is not needed because TypeScript supports ECMAScript 6
template strings, but in many cases they seem to uglify the code as
far as I'm concerned.

As such I've reverted all of the primary uses of such code to the old
format function, but I may use the new functionality in some places
going forward.

Lets say that this is an exercise in learning how to extend existing
classes in TypeScript.