Modify the blocking API to allow per actor blocking
This fulfills an issue in the issue tracker in which the API provided
by both the Tile and Entity class with regards to blocking Actor
instance movement across their bounds now take as a parameter an Actor
to test (which may be null).

The default implementations still always indicate blocking, but user
defined subclasses now have the granularity to say that some actors
can walk across a tile or enter the area of an entity while others
cannot, which might be interesting for future plans.