Clean accessors for the Point class
This class has been cleaned up to have all of its member variables
"warted" with a leading underscore and appropriate accessor methods
included. Member variables are now either private or protected as
needed.

In the case of a point, the X and Y values are private and are readable
but not  writable; in order to modify a point you need to use one of
the  provided methods.