Clean accessors for the Tile class
This class has been cleaned up to have all of its member variables
"warted" with a leading underscore and appropriate accessor methods
included. Member variables are now either private or protected as
needed.

In the case of a Tile, all members are protected because it is
supposed to be subclasses. The accessor methods for getting at the
name and id have different external names for the properties than
they have internally to make code that uses the tile a bit more
readable.

As a result, the Tileset class was also slightly modified so as to
use the new names.