Clean accessors for the Stage class
This class has been cleaned up to have all of its member variables
"warted" with a leading underscore and appropriate accessor methods
included. Member variables are now either private or protected as
needed.

In the case of the Stage, all members are private with accessors for
some of the needed values (some of which are still currently static).

This change necessitated some tweaks in both the Scene and Point
classes.