Clean accessors for the Scene class
This class has been cleaned up to have all of its member variables
"warted" with a leading underscore and appropriate accessor methods
included. Member variables are now either private or protected as
needed.

In the case of a Scene, the properties defined are all protected
because the class is designed to be sub-classed. The only exception is
the properties that are used as part of the screenshot creation code.
That code should not need to be touched in a subclass, so it can stay
hidden from accidental exposure.