Clean accessors for the LevelData class
This class has been cleaned up to have all of its member variables
"warted" with a leading underscore and appropriate accessor methods
included. Member variables are now either private or protected as
needed.

In the case of a LevelData, the member variables are protected as I
have in the past subclasses this class in order to do extra
post-processing on raw map data.

A part of this change alters the name of the external property that
contains the raw map data, which required a slight tweak to the Level
class as well.