Clean accessors for the Level class
This class has been cleaned up to have all of its member variables
"warted" with a leading underscore and appropriate accessor methods
included. Member variables are now either private or protected as
needed.

In the case of Level, much like LevelData I have made the properties
themselves be protected because some projects require a tweak to the
API that the Level provides to provide game specific functionality
that can't be folded into a base library.