Some project file tweaks
This includes some changes to the WebStorm project files and so are
not generally interesting.

The user dictionary for some words that I like to use that it doesn't
like have been updated.

Additionally I have turned off the compiler warning that a method in
a class can be static (determined by it not using any paramters or
manipulating instance state).

Much like the unused local or global symbol, I would love to leave
this turned on because it seems valuable, but for some reason you're
not allowed to insert a special comment that would disable it. In the
library there are a lot of cases of all three that are generating
warnings that are in error as far as I'm concerned.