Clean the Stage class a bit prior to refactoring
My idea for the engine is that you modify the code to set the stage
width and height by modifying the constants in the "header". This is
not really anything you want for production code, but for simple
prototyping it works well.

In the JavaScript version of the code I had the stage caching those
sizes in its own properties because otherwise everywhere that cares
about the width and height need to use nurdz.game.constants to get at
it, which is a huge drag.

In this version, since everything is in the same namespace it's a lot
cleaner to do it this way.