Refactor Scene handling from the Stage
There is now a SceneManager class whose job it is to hold all of the
scenes and switch between them, as well as determining which scene is
the current one.

This also cleans up some of the statics that used to exist in favor
of using instance functions instead, which makes things a little
cleaner.