Refactor Stage static variables
This cleans up the Stage class some more by factoring out the static
members that existed in the class in favor of just having them be
members in the module that don't get exported.

This captures them in the class closure and makes them static and
shared by all instances but makes references to them in code cleaner
by not requiring a "Stage." prefix.

The values still have an underscore prefix because they're local to
the module.

Note that this all falls on its face if for some reason the code is
ever modified to have multiple stages at once; more things than this
will break in that case, though. This ain't no multiplex.