Modify the Stage scene handling
The Stage makes sure that when a Scene switch is requested, it
doesn't happen until the end of the current update loop (which is
happening asynchronously) to avoid issues with updates happening for
one scene while frames are rendering for another.

This mechanism has been amended so that when a scene switch is
requested and the update loop is not currently running (usually
prior to the start of the game but also possibly at other points),
the scene switch happens right away.

This allows setup code to set a scene and switch to it, and then
modify the current scene and have it work as expected.