Enhance sample code
The sample main code is enhanced a bit. Now instead of just creating
a generic instance of Scene that does nothing it uses a small custom
version that draws the current FPS onto the stage.

Additionally it also contains a simple actor that represents a ball
that bounces itself around on the stage.

This allows us to make sure tha things are working (or mostly so),
anyway.