Modify the Point class API
In further thought, it is probably quite likely that we want to be
able to modify individual parts of the point itself.

As such, the values should not be hidden, so I have added a setter
for the X and Y values. Theoretically I could have just made the
values themselves public, but I wanted to keep the member variable
scheme of having the prepended underscore for internals.