Redact the stage setting interface on Actor
This is a holdover from a period of development on an old prototype
where entities did not have a stage set in their constructor due to
the way the level data was constructed. This has since been fixed.

It may be useful to do this in the future, but for right now the
ability to fiddle the stage out from under an Actor is gone.