Finish the Stage Rendering refactor
There is now a separate class that implements the Render interface
and does the magic. The Stage creates such an object from its canvas.

For now everything that needs to render needs to go through the Stage
class to its renderer property, which adds an extra level of
indirection.

This kind of bothers me from a clarity standpoint and should perhaps
be fixed. For now, this is good enough and that can be revisited later
if it appears to be a performance problem or too ugly to work with.