Change render function names
The render function names for drawing bitmaps have been renamed to blit
from what they were. This is only cosmetic, but it makes me feel better
because that's what I used to call it back in the day.

More importantly, I've wanted to do that for a while but I've never
gotten around to it and right now there are no calls to these methods
anywhere.