Modify how Stage reports FPS
In order to not blow anyone's brains, the Stage reports FPS by
calculating the value and then truncating it. This allows for
displaying the FPS without seeing a huge number of digits of precision.

This has been modified so that the FPS is rounded instead of
truncated. This probably loses a little precision but in a way that
stops a lot of FPS fluctuations caused by the clock skew of the
JavaScript timer granularity.