Include a Utilities API set
This is just a simple namespace (well not really, but yeah) with
exported functions that may come in handy in a game context anywhere.
As such they don't really fit with any class per se. I also didn't
want to generate these as static instances on a class, although the
code would undoubtedly be more or less the same.