New Entity Property system, part 1
The Entity class and the example have now been updated to how the
final Entity property system will operate. Unlike the JavaScript
version, here the Entity class takes default properties to apply as
part of the constructor.

Also, it is necesary for subclasses to  provide an overridden
_properties value to change the type to the one expected; otherwise,
the compiler thinks that it is still EntityProperties and will
complain about property usage. This is a bit of an annoyance but one
that I will put up with to get extra benefits from type completion
tools (problem could be solved by setting the _properties type to
"any", I think).

A bit more work still needs to be done before this is complete.