Final fix on the Entity property system
The last bits of this change fix up the entity validation system for
entity properties. The previous iteration would check that the type
of the property was valid, but that is now enforced by TypeScript, so
we don't have to.

The example code uses the new validation system to detect when the
speed of a dot is 0 on some axis so that it can be replaced with a
better value.