Modify the Rendering API
The rendering architecture as it currently existed was not originally
designed for the idea of anything scrolling anywhere (i.e a play area
larger than the stage in a virtual space); Actors (and thus entities)
always rendered at their current world position.

In order to facilitate future changes for this sort of thing, all of
the rendering API's take an explicit X,Y to render to. This allows
the Level to allow for scrolling by telling things to render in
positions where they currently are not.

At the same time I decided that since the Render API already had a
width and height value, the rendering methods should take it instead
of a Stage, so that was done.