Update the readme file in the engine branch
I'm contravening normal git best practices to keep different aspects
of devember separate from one another, and so I'm forking things all
over the place inside the repository instead of making multiple
repositories.

This is probably a maintenance nightmare but it keeps everything all
in one place for others and myself to peruse as far as devember
progress.