Fix a bug in the types used for input handling
TypeScript has different interfaces for the events used for keyboard
and mouse events that's different from general events, so that it can
type check better. The code was erroneously using a generic Event.

This only matters for type checking purposes; the generated code is
the same. As such it was easily worked around if you just implement
an overridden method with the correct type. However that's extra work.

This should have occurred to me during the initial port when I was
porting the code to convert mouse position and had to use MouseEvent
and not event.