Fix code formatting issues in a couple of places
WebStorm 11.0.2 came out today, and with it a fix to a regression in
their code formatter, which made it perform worse than it usually does.

In particular it was very bad about knowing that object.method() is a
function call, and that "name : type" in TypeScript is a type
expression in both method parameters as well as variable declarations.

Anyway, I was manually fixing some spacing while I was coding, and
since they fixed the bug (maybe, haven't checked too closely just
yet) and I was fixing a bug in the engine code anyway, I thought I
would go ahead and run the formatter over everything right now.