Allow the Stage/Scene to stop mouse event bubbling
The keyboard event handling in Stage and the Scene class have always
been set up so that the method has to return a boolean to indicate if
the event was handled, with the Stage stopping the event from being
further handled if yes.

This has now been extended to mouse events as well, because clicking
on things while debugging and having it select portions of the page
was annoying and I can't believe I never noticed this happening before.