Extend point API
Include the ability in the Point API to convert the point to an array
of two numbers and also to set a point from an array of (at least)
two values.

This is something that came up during the Rx prototyping during
devember day 10 as an easier way to construct polygon points, so it
might be useful elsewhere.