Implement a generic polygon renderer
This adds an interface to the rendering infrastructure that takes a
list of points (stored as arrays) and connects them together to plot
arbitrary polygons.

This represents points as arrays of numbers instead of actual Point
instances primarily because it's easiest to construct a polygon in
code with literal arrays as opposed to Point constructors.