Modify engine functions to return void
My bad on this one, I guess. It would appear that the default return
value for a function that doesn't have an explicit return type
specified and for which the code doesn't return a value is undefined
or something?

I've been having problems with WebStorm where overriding methods
would cause it to randomly decide if a method returns a value or not
as far as the signature inserted and the call to super.

It turns out that this is because the compiler doesn't know for sure
that the methods are not supposed to return anything.

As such, this is now fixed by explicitly declaring the return type
for everything in the engine API.