Enhance Renderer API
The Rendering API in Renderer and CanvasRenderer has been enhanced a
bit.

Firstly, the polygon API has been modified to allow for sub shapes.
Now instead of a list of points, the points can optionally contain a
coommand string to indicate that the point represents a lineTo,
moveTo or closePath, so that a polygon cn be more complex.

Secondly, all of the shape API's now have a stroke version as well,
so that you can create outlines if desired and not just fills.

C), the Actor now uses a strokeRect call instead of a fillRect() call
for default debug rendering, which makes things less busy.