Include global tick count in the engine
The stage now keeps a count of the number of update ticks that have
happened, and this is passed to the scene class in its update method
as well as to actors in their update method.

This allows for a crude (and stupid) way for things to be timed out
in the engine. It's stupid because it's tied to the update rate and
not the frame rate, which seems wrong.

I'll worry about this in the future; probably the actual delta since
the last frame should be used instead, or the update loop should
always run at a constant tick rate independent of the render loop
rate or something.