Shift the logic for taking a screenshot from Scene to Stage
This seems perhaps better placed here because the Stage is what owns the canvas,
which is needed for screen shot capturing.

As a part of this change, there are static variables in Stage for setting the
default filename and window title used by the screen shot system, so that you
can just call the screenshot method directly.

The sample application sets the values to what the old defaults used to be.