Fix bug in CanvasRenderer exposed by new image preloading
Up until now I have not tested the image blitting portion of the new
TypeScript engine, and so I was unaware that the code for blitting an
image centered on a location was broken.

Specifically it would translate the canvas to the drawing location,
but then never restore the context. As a result all of the images
render at the same location always. This is now fixed.