Enhance image preload to allow image reuse
Now that the basic preload is working, the obvious extension to what
it's doing is to allow for multiple requests for the same image to
actually only load the same image once.

The browser would reuse the underlying data anyway, but we can (I
think?) save some object instances by just detecting when an attempt
to preload something that someone already asked for a preload for and
return the same image tag that the first one returned.