Include (primitive) sound preloading as well
This uses the same pattern as the image loading does, but loads
sounds instead of images.

This still needs more work (for example, a Sound and Music class or
something) because just invoking the .play() method will do nothing
if the sound is already playing, so if multiple things have the same
sound they can't all trigger at once.

I'm not sure yet what the best solution is to that particular issue,
but at least we can annoy people with bloops now!