Include some music files
An upcoming commit is going to include the ability to preload and play
music, so it's added now.

The song included:
"Who Likes to Party" Kevin MacLeod (incompetech.com)
Licensed under Creative Commons: By Attribution 3.0
http://creativecommons.org/licenses/by/3.0/

I pulled down the mp3 version, turned it into a wav and then
converted the wav to an ogg file.