Load music in the demo "game"
We load the included music in the demo game and then start it playing
when everything starts.

There is currently an issue with this in that you can switch scenes
while the game is not currently running, but doing so triggers the
appropriate callbacks, which happen before the preload begins, so
anything in those methods that rely on pre-loaded content will fail.

This needs some looking into.