Allow for music and sound muting
Let there be the ability to stop music from playing in the background so
that it doesn't drive you bananas while you're busy tweaking code.

This (kind of crudely) also mutes the sounds that play by adding a
property to the dot objects that indicates if they should play their
sound effect when they bounce or not, which is turned on by default
but tweaked to track the playing state of the music.

A better way to do this would be to have the Sound (and thus also
Music) class have a static method that makes the play method do
nothing when a flag is set (global mute), but this would require the
class to track all of its own instances, or at least music instances,
or toggling the state would not fully work.

That's outside scope for right now.

(Note: I guess it could also just exist and it's up to you to make
sure that you pause any playing music and sounds would take care of
themselves).