Change sound preload to not share tags
With the previous commit it seems important that unlike images,
preloaded sounds don't short circuit and return back to you the
previous tag that was returned when the same sound is loaded more
than once.

In doing this changing the volume on one would affect all (for
example), which is not desirable.