Split Engine and Game into separate JS files
Phase one of this is now complete; there are two tsconfig.json files
to compile each part to its own output file and both are included
in the HTML file.

I'm currently not doing anything with external build systems of any
kind (maybe next?) so the .d.ts file falls out inside the JS output
directory because the TypeScript compiler puts both in the same place
no matter what.

As a result, this means that the tsconfig file for the game has to
reference the file in that location, which is a little bit of a drag,
but oh well.

In theory if one was to have the engine be its own project, this
would be taken care of by not having the compiler generate those
files anyway.