Catch errors in the preloader code
We now catch errors in the preload using appropriate event handlers.
In the case that an image or sound cannot be loaded, a version from a
data URL embedded in the code is used instead, so that things
continue to work.

The replacement image is a red X in a box while the audio is just
short silence.