Include screenshot/thumbnail image
As part of the new upcoming web site, projects using this template
should provide a thumbnail image that can be used as part of the link
to their page. This should be maintained as a part of the actual
project and not a part of the website that hosts it.