Include a proper brand link
Since there will (hopefully) be multiple pages using this particular
template, there should be a proper link back to the main site.

This pushes the previous "brand" (which was a project name) to be
some static text in the navbar instead.

Side Note: Does anyone out there know what magic voodoo IntelliJ
products use to determine when they should add words to the per
project dictionary and when they should be packed into something in
the actual install? Because it seems really random to be and entirely
irritating in almost every possible way.