Include callbacks on a per image/sound/music basis for Preload
Now besides just having a single "all preloads are now done" callback
(which the stage uses to know it's time to start the game), things
that ask for a preload can specify that a callback be invoked when
the preload is completed.

This means that things that need to know about the dimensions of an
image can be told as soon as its done loading, for example.