Include an initial SpriteSheet class
We now have an initial simple SpriteSheet class. This assumes that an
image is a sheet of tiles of an identical size. The sheet can be
created either by indicating how many sprites across and down the
image contains or by indicating the dimensions of each sprite.

A simple rendering facility is included which performs the same
rendering as can be done for regular images so that single sprites
can be rendered in any way.