Include an origin property on Actor
This specifies an offset from the top left that the Actor's position
should be relative to for purposes of rendering and collision
detection.

To this end, the Scene now applies the origin to the location that it
gives to the Actor.render() method, so that Actor.render() should
always assume that it is rendering at the upper left corner.

This removes a lot of the complications of some Actors having a
position that's not relative to a known location (something that came
up in ts-tennis).