Include a debug property in Entity
This adds an extra default property to all entities named debug. This
can be used in any way desired by your code, to allow for debugging
to only occur for certain entities or classes of entity.

The Entity base class now has a render method which checks to see if
the debug property is set, and if it is, it invokes the Actor render()
method. That method renders the bounding box and origin of the Actor,
so this provides an easy way to track positioning of things if
desired.