Update the sample to use new origins and debug code
The Dot class in the sample now renders itself properly, since the
meaning of the coordinates given to the render method have changed.
The method also invokes the super class to take advantage of debugging.

One of the two balls has debug mode turned on just to illustrate now
that works.