Give Actor/Entity inherent sprite rendering abilities
The API is enhanced a bit so that the Actor class (and thus its
subclass Entity) have a _sheet and _sprite private property. The
first is a sprite sheet, which defaults to null, and the second is a
numeric index into the sprite sheet, which defaults to 0.

If an Actor has a sprite sheet attached AND the sprite index is valid
for it, then the default render() method renders the sprite. If there
is no sprite sheet attached or the sprite index is invalid, do what
we did before and render bounds instead.

Similarly, the Entity class now invokes the super method always, so
it always gets either the current sprite or a bounding box if there
isn't one. Additionally, if the debug property is turned on and the
Actor class didn't already render bounds, then the Entity class
renders that as well.

With these changes, most entities will never need to even have their
own render method, which is nice.