Include Actor/Entity default rotation for display
This adds a property to Actor named angle which represents the angle
to rotate the canvas to prior to doing any rendering for this
particular entity.

In practice, this translates the canvas to the origin location,
rotates it, renders, and then restores the canvas.

This could perhaps be done more intelligently; the render method is
getting a location value that is adjusted for the origin offset so
that it knows it always has to render from the upper left, but the
code in the render method now immediately undoes that calculation.

That's something to look at tomorrow, but for now I'm checking this
in (along with a change to test it) and then it can be looked at to
make sure things don't break.