Optimize rendering a bit
As mentioned in a previous commit, the rendering pipeline has been
optimized slightly. Previously the Scene class would apply the origin
to the position when it invoked the render method, and the Actor and
Entity methods would promptly undo this.

Since I expect that most of the time rendering will be done by this
code and not with a custom renderer, it seems better to have the render
code take on the task of handling the origin if needed.