New Stage API : get tick speed
This allows you to query the tick speed of the engine while it is
running, allowing you to determine how many ticks a second is, while
allows for proper timing.

This only provides a value while the engine is actively running,
which means that the value reported is 0 right up until the run()
method is called AND the preload has completed.