Implement an initial Animation controller class
The class is named AnimationList, and is a simple bit of code that
wraps a list of frames and a rate of play and associates such lists
with textual names.

There is an update function which should be invoked at each game tick
and determines when the frame in the animation should change. This
takes care of looping and ping pong as needed.

The function always returns 0 when there is no animation to play, so
possibly you may want to not invoke it if you plan on stopping
animation playing at any point.