Implement a proxy for the animation code in Actor
The Actor class now has its own animation list, which it lazily
initializes as needed.

The API of the animation list is proxied into the Actor class such
that it appears to be a native API; the first time an animation is
added, the animation list is created.

This change also fixes some comment issues in the animation list that
slipped through the last commit.