Tweak the animation API exposed in Actor
Instead of having the entirety of the AnimationList API in Actor, we
instead expose just the functions for adding and playing an
animation, and expose the animation list property for more complete
actions.

This cleans up the code while making it handy for the main useful
functions to be invoked directly on Actor.