Mouse position function now takes an optional parameter
The Stage.calculateMousePos() method that tells you where on the
stage a mouse event happened has been modified. Previously every time
you invoked it, it would create a new Point object with the mouse
location and return it.

Now it takes an optional parameter which is a Point; If this is
given, the point is modified and returned, otherwise the old
behaviour is used.

This can be used to curtail the creation of multiple points, if
desired, since the mouse loves a lot and this could introduce a lot
of garbage.