Change the Dot entities to have circular collision type
As an initial test for the system, we modify at construction time the
collision type of the dot entities to be circular instead of
rectangular.

There will be an official API for making this change, but for the
time being we're doing this manually just to ensure that everything
is still working as expected.