Include some collision code
This tweaks the Utils function library to include some extra methods,
plus a new module that contains core collision detection routines.
These allow for hit detection of a point in circles and rectangles,
and intersections between circles and rectangles (either
circle/circle, rectangle/rectangle or circle/rectangle).

Lastly, there are some collision methods for detecting the
intersection between two lines or line segments.

This is not currently used for anything in the engine, that's the
subject of upcoming commits.