Start the merge of the Actor and Collider class
This was the intention at the outset; all Actors should be colliders,
so that they can collide/be hit tested/etc but at the same time we
could just make random colliders to check for things (e.g. triggers).

This removes the redundant properties from the Actor class and fixes
up the rendering of the Actor and Entity class (the name of the
method that renders the bounding box is changed).

Currently all actors are inherently rectangles and there is no way to
set a non-circular collision type except to know how to do so manually.
The example code works as always using the new code but a bit more
development is needed.