Include the new Collider class
This is the class developed in isolate in a private branch; I've
moved it here in whole instead of as a series of commits due to the
long and circuitous route it took to come up with this design.

This class is meant to be a common ancestor for the Actor class as
well as other classes that might need to be used for collisions (e.g.
simple in-game collision volumes or such) and so it has all of the
common properties that an Actor would need.

The idea is that each collider has an inherent type and volume
properties (width/height or radius) and supports an API that allows
for checking for a collision via a point in or with another collider.

Colliders may also have the NONE type to indicate that they exist but
do not collide, which might be useful for simple display purposes,
for example.