Include the ability to detect when the mouse is pressed or released
This happens in concert with the click event, which has been modified
internally to actually represent a click event (it was named properly
but actually tracked a mouse press and not a click).

Additionally, although the click and mouse down events only trigger
when the mouse is over the canvas, the mouse up (button released)
event triggers everywhere. This allows the detection of the mouse
button being released if its pressed inside the stage and then
dragged out prior to release.