Get ready for line->shape collisions
An extra set of points has been added to render a line for checking
collisions with shapes. These points are stored the same as the other
points, but the line color is changed before rendering the line.

Additionally, we no longer allow the shape/point collision involving
the shape that follows the mouse to collide with the static shapes on
the stage while doing a mouse drag.