Include a simple Vector class
This represents only 2D vectors (all we care about at the moment) and
has a bit of an API. Some changes could probably be done here.

For one thing, the Point class and this class are very similar in
many respects, so it might be a good idea to have a common ancestor?

Also, the Point class could be extended to allow moving a point along
a vector or such; we'll leave that to when it's actually needed,
though.