Enhance the Point/Vector classes
The Point class gains a static construction method that can create a
point given a vector and an optional origin, the ability to calculate
the rotation between itself and another point and the distance between
itself and another point.

The Vector class now allows you to set its magnitude as a simple wrapper
around first normalizing and then scaling. We also modify the method
that returns a copy of the vector after it is made orthogonal to itself
to modify the vector directly and then make a new copy method. This
covers the use case of wanting to just rotate a vector without
creating a new one.