Enhance utilities
Based on the previous edits, add in some extra utility methods for
finding the distance between two points as the square of the distance
instead of the actual distance.

These functions are used in the circle collision routines, so those
have been modified to compare the square of the distance to the
square of the radius under the theory that a multiplication is faster
than a square root for the purposes of a comparison that doesn't
actually return the actual distance.