More enhanced utility code
Include the ability to normalize a value within a range and do a
linear interpolation from a normalized value to a range.

Both of these operations are simple, but having methods for them
helps to document what they're being used for, so I decided that they
should exist anyway. This eases the inner tension of having repeated
code that could be factored out.