Tighten up the code a bit
This change does two things; firstly it adds in the notion of
changing the direction of the vector by setting a property in the
same way that setting the magnitude can be done the same way.

This just makes the API look nicer, and some minor code changes need
to occur where fetching the current direction is now no longer a
method call.

Secondly, the code to alter the magnitude is done via trig instead of
the simple way I originally had, which was to normalize the vector
and then scale it.

This is done on the presumption that calculating sine and cosine are
faster than a square root. This may or may not actually be the case,
though.