Include Line/Rect intersect and fix formatting
This includes a utility collision routine for determining the collision
point between a rectangle and a line segment. The methods will tell
you the exact point that the segment intersects with the rectangle,
using the directionality of the line segment to determine which of
multiple intersections is the correct one.

Additionally, I didn't realize that WebStorm was a tit about indenting
the code due to there being a leading comment (yet another instance of
broken code formatting, sigh), so that is fixed as well.