Include tests for line/shape collisions
This includes the formal testing for the new collision code. This is
set up to try to collide with all shapes and not just the rectangles,
so that when we include that ability (eventually) we will be ready to
go.

The control line for this now renders with an arrow head only on its
end, allowing us to visuzlize it's direction better. Additionally,
the perpendicular is rendered at the originating endpoint of the line.

This is not actually needed for anything, but it's a nice visualization
of which side of the line is the left side and that the vector code
is working.